Refactored Home page
